test: add test cases for runtime read
